fdserver: require contexts to be created
Contexts will have to be created by the user before registering file
descriptors or doing any other operations with the server (except
for now stopping it).

Until now contexts had to be predefined in fdserver_context_e, but with
this work this will be moved to an opaque type, and each client will be
able to register different contexts at will without having to have them
predefined.

Signed-off-by: Josep Puigdemont <josep.puigdemont@linaro.org>