fdserver_internals: use opaque context internally
When creating a new context, the server now provides a tuple that
identifies it, and that the client must use for further communication.
This is another step towards making the context type opaque.

The tuple consists of a index identifying a position in an array
(currently it only has one element), and a random token (currently a
well known one), both must match in order to identify a "connection".

Signed-off-by: Josep Puigdemont <josep.puigdemont@linaro.org>