fdserver: make socket path configurable
The server can be called with "-p /path/to/socket", and it will create
the socket on the specified path (checks for validity are missing atm)

The clients should now call fdserver_init(path), to configure the path
the library should use. The user can pass NULL to use the default path.

Signed-off-by: Josep Puigdemont <josep.puigdemont@linaro.org>